restructure folder hiearchy
