+++ frontend/src/app/modules/dyad/models/dyad.models.ts
