+++ frontend/src/app/modules/precise/models/participant-summary.model.ts
