+++ frontend/src/app/modules/shared/animations.ts
