+++ frontend/src/app/modules/shared/components/common/notificationBar.ts
