+++ frontend/src/app/modules/shared/components/icons.ts
