+++ frontend/src/app/modules/shared/components/odk/odkTableRows/odkTableRows.ts
