+++ frontend/src/app/modules/shared/pages/developer-tools/developer-tools.component.spec.ts
