+++ frontend/src/app/modules/shared/pages/home/home.component.scss
