+++ frontend/src/app/modules/shared/pages/install/install.component.spec.ts
