+++ frontend/src/app/modules/shared/pages/install/install.component.ts
