+++ frontend/src/app/modules/shared/pipes/boolIconPipe.ts
