+++ frontend/src/app/modules/shared/pipes/index.ts
