+++ frontend/src/app/modules/shared/services/odk/altMockImplementation/odkCommon.ts
