+++ frontend/src/app/modules/shared/services/utils.ts
