+++ frontend/src/app/modules/shared/stores/common.store.ts
