+++ frontend/src/app/modules/shared/stores/index.ts
@@ -1,0 +1,1 @@
+export * from "./common.store";
