+++ frontend/src/app/modules/shared/utils/guid.ts
