+++ frontend/src/app/modules/shared/utils/index.ts
