--- frontend/src/app/stores/index.ts
@@ -1,2 +1,0 @@
-export * from "./common.store";
-export * from "./precise.store";
